update puppetlabs release rpm path construction for EL9
The http://yum.puppet.com/puppet7 directory does not contain an rpm for
EL9. In other words, this URL constructed using the current logic is a
404:

  https://yum.puppet.com/puppet7/puppet7-release-el-9.noarch.rpm

There are packages for EL5 through EL9 in the root of the repo. E.g.:

  http://yum.puppet.com/puppet7-release-el-9.noarch.rpm

This simple fix is to drop usage of a "puppet<ELX>" subdir for all major
EL versions.